DOCUMENT: Introduce the concept of special discount
Before applying this commit, product has only one discount, i.e. the
current discount. The problem of this design is we cannot set different
discount for a product in different catalog.

This commit introduce special discount of a product, which can be
different for each catalog. For example, a book can have current
discount of 30%, special discount of 40% in Deepavali catalog, and
special discount of 50% in Christmas catalog.